refactor(nuxt): keep bridge additions within source length gate
Split the Nuxt keyed-functions plugin patch and the npmx compare mocks into
dedicated modules, move the new bridge tests out of index.test.ts, and shorten
the og-image SFC exclude so no file over 350 lines grows in this PR.
