+++ npm/framework/nuxt/src/index.ts
@@ -6,6 +6,7 @@ import { appendMuseaArtComponentIgnore } from "./musea-components";
 import { registerNuxtMuseaStaticPublicAsset } from "./musea-static";
 import { patchNuxtClientManifestCloseBundlePlugin } from "./client-manifest-bridge";
 import { patchNuxtHostVuePluginForCompilerExcludes } from "./host-vue-bridge";
+import { patchNuxtKeyedFunctionsPlugin, type ViteTransformResult } from "./keyed-functions-bridge";
 import "./schema";
 import * as bridgeFastPath from "./bridge-fast-path";
 import type { VizeNuxtCompilerOptions, VizeNuxtOptions } from "./options";
@@ -28,7 +29,6 @@ import {
   stabilizeNuxtInjectedKeysForVizeVirtualModule,
 } from "./utils";
 import { appendOriginalVueSourceForUnoCss } from "./unocss";
-type ViteTransformResult = string | { code?: string; map?: unknown } | null | undefined;
 const VIZE_NUXT_AUTO_IMPORT_PATCHED = "__vizeNuxtAutoImportPatched";
 const VUE_RUNTIME_DEDUPE = [
   "vue",
@@ -209,50 +209,6 @@ function isViteSsrTransform(args: unknown[]): boolean {
     "ssr" in options &&
     (options as { ssr?: boolean }).ssr === true
   );
-}
-
-function normalizeNuxtKeyedTransformResult(
-  id: string,
-  result: ViteTransformResult,
-): ViteTransformResult {
-  if (!isVizeVirtualVueModuleId(id) || result == null) {
-    return result;
-  }
-  if (typeof result === "string") {
-    return normalizeNuxtInjectedKeysForVizeVirtualModule(result, id);
-  }
-  if (typeof result.code !== "string") {
-    return result;
-  }
-  const code = normalizeNuxtInjectedKeysForVizeVirtualModule(result.code, id);
-  return code === result.code ? result : { ...result, code };
-}
-
-function patchNuxtKeyedFunctionsPlugin(plugin: { transform?: unknown }): void {
-  if (typeof plugin.transform === "function") {
-    const original = plugin.transform;
-    plugin.transform = async function (
-      this: unknown,
-      code: string,
-      id: string,
-      ...args: unknown[]
-    ) {
-      const result = (await original.call(this, code, id, ...args)) as ViteTransformResult;
-      return normalizeNuxtKeyedTransformResult(id, result);
-    };
-    return;
-  }
-
-  const transform = plugin.transform as { handler?: unknown } | undefined;
-  if (!transform || typeof transform.handler !== "function") {
-    return;
-  }
-
-  const original = transform.handler;
-  transform.handler = async function (this: unknown, code: string, id: string, ...args: unknown[]) {
-    const result = (await original.call(this, code, id, ...args)) as ViteTransformResult;
-    return normalizeNuxtKeyedTransformResult(id, result);
-  };
 }
 
 function normalizeNuxtAutoImportTransformResult(
